requirements.txt
azure-devops==6.0.0b4
certifi==2022.9.24
charset-normalizer==2.1.1
factory-boy==3.2.1
Faker==15.1.1
idna==3.4
isodate==0.6.1
msrest==0.6.21
oauthlib==3.2.2
python-dateutil==2.8.2
requests==2.28.1
requests-oauthlib==1.3.1
six==1.16.0
urllib3==1.26.12
